+++ gettsim/__init__.py
@@ -5,18 +5,18 @@ import pytest
 
 from gettsim.config import ROOT_DIR
 from gettsim.interface import compute_taxes_and_transfers  # noqa: F401
-from gettsim.pre_processing.policy_for_date import get_policies_for_date  # noqa: F401
+from gettsim.policy_environment import set_up_policy_environment  # noqa: F401
 
 __version__ = "0.3.3"
 
 COUNTER_TEST_EXECUTIONS = itertools.count()
 
 
-def test():
+def test(*args):
     n_test_executions = next(COUNTER_TEST_EXECUTIONS)
 
     if n_test_executions == 0:
-        pytest.main([str(ROOT_DIR)])
+        pytest.main([str(ROOT_DIR), *args])
     else:
         warnings.warn(
             "Repeated execution of the test suite is not possible. Start a new Python"