+++ gettsim/functions_loader.py
@@ -2,16 +2,27 @@ import importlib
 import inspect
 from pathlib import Path
 
+from gettsim.config import PATHS_TO_INTERNAL_FUNCTIONS
 from gettsim.config import ROOT_DIR
 
 
-def convert_paths_to_import_strings(paths):
+def load_user_and_internal_functions(functions):
+    functions = [] if functions is None else functions
+
+    custom_functions = _load_functions(functions)
+    imports = _convert_paths_to_import_strings(PATHS_TO_INTERNAL_FUNCTIONS)
+    internal_functions = _load_functions(imports)
+
+    return custom_functions, internal_functions
+
+
+def _convert_paths_to_import_strings(paths):
     """Convert paths to modules for gettsim's internal functions to imports.
 
     Example
     -------
     >>> path = ROOT_DIR / "demographic_vars.py"
-    >>> convert_paths_to_import_strings(path)
+    >>> _convert_paths_to_import_strings(path)
     ['gettsim.demographic_vars']
 
     """
@@ -23,12 +34,12 @@ def convert_paths_to_import_strings(paths):
     return import_strings
 
 
-def load_functions(sources, allow_imported_members=False):
+def _load_functions(sources, allow_imported_members=False):
     """Load functions.
 
     Parameters
     ----------
-    sources : str, pathlib.Path, function, module, dictionary of functions
+    sources : str, pathlib.Path, function, module, imports statements
         Sources from where to load functions.
     allow_imported_members : bool
         Should imported members also be collected from a module?
@@ -41,7 +52,6 @@ def load_functions(sources, allow_imported_members=False):
 
     """
     sources = sources if isinstance(sources, list) else [sources]
-    sources = _convert_some_strings_to_paths(sources)
     sources = _search_directories_recursively_for_python_files(sources)
     sources = _convert_paths_and_strings_to_dicts_of_functions(
         sources, allow_imported_members
@@ -55,16 +65,6 @@ def load_functions(sources, allow_imported_members=False):
         if isinstance(source, dict) and all(
             inspect.isfunction(i) for i in source.values()
         ):
-            # Test whether there are duplicate functions.
-            duplicated_functions = set(source) & set(functions)
-            if duplicated_functions and not allow_imported_members:
-                formatted = _format_duplicated_functions(
-                    duplicated_functions, functions, source
-                )
-                raise ValueError(
-                    f"The following functions are defined multiple times:\n{formatted}"
-                )
-
             functions = {**functions, **source}
 
         else:
@@ -73,31 +73,6 @@ def load_functions(sources, allow_imported_members=False):
             )
 
     return functions
-
-
-def _convert_some_strings_to_paths(sources):
-    """Handle strings in sources.
-
-    Strings are evaluated by :func:`importlib.utils.find_spec` to see whether it is a
-    module import. If an error occurs or it returns None, the string is converted to a
-    path.
-
-    """
-    new_sources = []
-    for source in sources:
-        if isinstance(source, str):
-            try:
-                out = importlib.util.find_spec(source)
-            except ValueError:
-                out = Path(source)
-            else:
-                out = Path(source) if out is None else source
-        else:
-            out = source
-
-        new_sources.append(source)
-
-    return new_sources
 
 
 def _search_directories_recursively_for_python_files(sources):