+++ gettsim/tests/test_arbeitsl_geld.py
@@ -4,7 +4,7 @@ from pandas.testing import assert_series_equal
 
 from gettsim.config import ROOT_DIR
 from gettsim.interface import compute_taxes_and_transfers
-from gettsim.pre_processing.policy_for_date import get_policies_for_date
+from gettsim.policy_environment import set_up_policy_environment
 
 INPUT_COLS = [
     "p_id",
@@ -37,19 +37,13 @@ def test_ui(
 ):
     year_data = input_data[input_data["jahr"] == year]
     df = year_data[INPUT_COLS].copy()
-    params_dict, policy_func_dict = get_policies_for_date(
-        policy_date=year,
-        policy_groups=[
-            "arbeitsl_geld",
-            "soz_vers_beitr",
-            "eink_st_abzuege",
-            "eink_st",
-            "soli_st",
-        ],
-    )
+    policy_params, policy_functions = set_up_policy_environment(date=year)
 
     result = compute_taxes_and_transfers(
-        df, targets="arbeitsl_geld_m", params=params_dict
+        data=df,
+        params=policy_params,
+        functions=policy_functions,
+        targets="arbeitsl_geld_m",
     )
 
     # to prevent errors from rounding, allow deviations after the 3rd digit.