+++ gettsim/tests/test_elterngeld.py
@@ -6,7 +6,7 @@ from pandas.testing import assert_series_equal
 
 from gettsim.config import ROOT_DIR
 from gettsim.interface import compute_taxes_and_transfers
-from gettsim.pre_processing.policy_for_date import get_policies_for_date
+from gettsim.policy_environment import set_up_policy_environment
 
 INPUT_COLS = [
     "hh_id",
@@ -58,23 +58,18 @@ def test_eltgeld(
     """
     year_data = input_data[input_data["jahr"] == year]
     df = year_data[INPUT_COLS].copy()
-    params_dict, policy_func_dict = get_policies_for_date(
-        policy_date=year,
-        policy_groups=[
-            "elterngeld",
-            "soz_vers_beitr",
-            "eink_st_abzuege",
-            "eink_st",
-            "soli_st",
-        ],
-    )
+    policy_params, policy_functions = set_up_policy_environment(date=year)
     df["soli_st_tu"] = df["soli_st_m"].groupby(df["tu_id"]).transform("sum") * 12
     df["eink_st_tu"] = df["eink_st_m"].groupby(df["tu_id"]).transform("sum") * 12
 
-    columns = ["soli_st_tu", "sozialv_beitr_m"]
+    columns_overriding_functions = ["soli_st_tu", "sozialv_beitr_m", "eink_st_tu"]
 
     result = compute_taxes_and_transfers(
-        df, user_columns=columns, targets=column, params=params_dict
+        data=df,
+        params=policy_params,
+        functions=policy_functions,
+        targets=column,
+        columns_overriding_functions=columns_overriding_functions,
     )
 
     assert_series_equal(