+++ gettsim/tests/test_functions_loader.py
@@ -1,9 +1,7 @@
 import textwrap
 from pathlib import Path
 
-import pytest
-
-from gettsim.functions_loader import load_functions
+from gettsim.functions_loader import _load_functions
 
 
 def func():
@@ -11,26 +9,21 @@ def func():
 
 
 def test_load_function():
-    assert load_functions(func) == {"func": func}
-
-
-def test_raise_error_on_duplicated_functions():
-    with pytest.raises(ValueError):
-        load_functions([func, func])
+    assert _load_functions(func) == {"func": func}
 
 
 def test_renaming_functions():
-    out = load_functions([func, {"func_": func}])
+    out = _load_functions([func, {"func_": func}])
     assert len(out) == 2
     assert "func" in out and "func_" in out
 
 
 def test_load_modules():
-    assert load_functions("gettsim.soz_vers.krankenv_pflegev")
+    assert _load_functions("gettsim.soz_vers.krankenv_pflegev")
 
 
 def test_load_path():
-    assert load_functions(
+    assert _load_functions(
         Path(__file__).parent.joinpath("..", "soz_vers", "krankenv_pflegev.py")
     )
 
@@ -42,7 +35,7 @@ def test_special_attribute_module_is_set(tmp_path):
     """
     tmp_path.joinpath("functions.py").write_text(textwrap.dedent(py_file))
 
-    out = load_functions(tmp_path.joinpath("functions.py"))
+    out = _load_functions(tmp_path.joinpath("functions.py"))
     assert isinstance(out, dict)
     assert "func" in out
     assert len(out) == 1
@@ -50,7 +43,7 @@ def test_special_attribute_module_is_set(tmp_path):
 
 
 def test_special_attribute_module_is_set_for_internal_functions():
-    out = load_functions("gettsim.soz_vers.eink_grenzen")
+    out = _load_functions("gettsim.soz_vers.eink_grenzen")
     function = out[list(out)[0]]
 
     assert function.__module__ == "gettsim.soz_vers.eink_grenzen"