+++ gettsim/tests/test_wohngeld.py
@@ -7,7 +7,7 @@ from pandas.testing import assert_series_equal
 
 from gettsim.config import ROOT_DIR
 from gettsim.interface import compute_taxes_and_transfers
-from gettsim.pre_processing.policy_for_date import get_policies_for_date
+from gettsim.policy_environment import set_up_policy_environment
 
 
 # Variables for the standard wohngeld test.
@@ -57,9 +57,7 @@ def input_data():
 def test_wg(input_data, year, column):
     year_data = input_data[input_data["jahr"] == year]
     df = year_data[INPUT_COLS].copy()
-    params_dict, policy_func_dict = get_policies_for_date(
-        policy_date=year, policy_groups="wohngeld"
-    )
+    policy_params, policy_functions = set_up_policy_environment(date=year)
     columns = [
         "elterngeld_m",
         "arbeitsl_geld_m",
@@ -73,14 +71,14 @@ def test_wg(input_data, year, column):
         "rentenv_beitr_m",
         "kindergeld_anspruch",
     ]
-    policy_func_dict["eink_st_tu"] = eink_st_m_tu_from_data
+    policy_functions["eink_st_tu"] = eink_st_m_tu_from_data
 
     result = compute_taxes_and_transfers(
-        df,
-        user_columns=columns,
-        user_functions=policy_func_dict,
+        data=df,
+        params=policy_params,
+        functions=policy_functions,
         targets=column,
-        params=params_dict,
+        columns_overriding_functions=columns,
     )
     assert_series_equal(result[column], year_data[column])
 
@@ -128,9 +126,7 @@ def input_data_households():
 @pytest.mark.parametrize("year, mietstufe", itertools.product(POLICY_YEARS, MIETSTUFEN))
 def test_increasing_hh_size(input_data_households, year, mietstufe):
     column = "wohngeld_basis_hh"
-    params_dict, policy_func_dict = get_policies_for_date(
-        policy_date=year, policy_groups="wohngeld"
-    )
+    policy_params, policy_functions = set_up_policy_environment(date=year)
     columns = [
         "elterngeld_m",
         "arbeitsl_geld_m",
@@ -148,10 +144,10 @@ def test_increasing_hh_size(input_data_households, year, mietstufe):
     input_data_households["mietstufe"] = mietstufe
 
     result = compute_taxes_and_transfers(
-        input_data_households,
-        user_columns=columns,
-        user_functions=policy_func_dict,
+        data=input_data_households,
+        params=policy_params,
+        functions=policy_functions,
         targets=column,
-        params=params_dict,
+        columns_overriding_functions=columns,
     )
     assert result[column].is_monotonic